AI-Agent/backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# API Clients
requests==2.31.0
google-genai==0.3.0

# Vector Database
pinecone-client==4.1.2

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
